flask_app/requirements.txt
flask-bootstrap
wtforms
gunicorn
flask-appconfig>=0.10
flask-nav
flask-debug
flask-wtf
itsdangerous==2.0.1